working out the authentication
