+++ piJEE/piJEE-ejb/src/main/java/tn/esprit/overpowered/byusforus/entities/authentication/Session.java
@@ -1,0 +1,69 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package tn.esprit.overpowered.byusforus.entities.authentication;
+
+import java.io.Serializable;
+import java.time.LocalDateTime;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.OneToOne;
+import tn.esprit.overpowered.byusforus.entities.users.User;
+
+/**
+ *
+ * @author EliteBook
+ */
+@Entity
+public class Session implements Serializable {
+
+    private static final long serialVersionUID = 1L;
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
+    
+    @OneToOne
+    private User user;
+    
+    private LocalDateTime started;
+    private int expiresAfter;
+    private String clientType;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    @Override
+    public int hashCode() {
+        int hash = 0;
+        hash += (id != null ? id.hashCode() : 0);
+        return hash;
+    }
+
+    @Override
+    public boolean equals(Object object) {
+        // TODO: Warning - this method won't work in the case the id fields are not set
+        if (!(object instanceof Session)) {
+            return false;
+        }
+        Session other = (Session) object;
+        if ((this.id == null && other.id != null) || (this.id != null && !this.id.equals(other.id))) {
+            return false;
+        }
+        return true;
+    }
+
+    @Override
+    public String toString() {
+        return "tn.esprit.overpowered.byusforus.entities.authentication.Session[ id=" + id + " ]";
+    }
+    
+}
