--- piJEE/piJEE-ejb/src/main/java/tn/esprit/overpowered/byusforus/entities/entrepriseprofile/ProjectManager.java
@@ -1,61 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package tn.esprit.overpowered.byusforus.entities.entrepriseprofile;
-
-import java.io.Serializable;
-import javax.persistence.DiscriminatorValue;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-/**
- *
- * @author pc
- */
-@Entity
-@DiscriminatorValue(value = "PROJ_MANAGER")
-public class ProjectManager extends Employee implements Serializable {
-
-    private static final long serialVersionUID = 1L;
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    @Override
-    public int hashCode() {
-        int hash = 0;
-        hash += (id != null ? id.hashCode() : 0);
-        return hash;
-    }
-
-    @Override
-    public boolean equals(Object object) {
-        // TODO: Warning - this method won't work in the case the id fields are not set
-        if (!(object instanceof ProjectManager)) {
-            return false;
-        }
-        ProjectManager other = (ProjectManager) object;
-        if ((this.id == null && other.id != null) || (this.id != null && !this.id.equals(other.id))) {
-            return false;
-        }
-        return true;
-    }
-
-    @Override
-    public String toString() {
-        return "tn.esprit.overpowered.byusforus.entities.ProjectManager[ id=" + id + " ]";
-    }
-    
-}
