+++ piJEE/piJEE-ejb/src/main/java/tn/esprit/overpowered/byusforus/entities/users/User.java
@@ -3,9 +3,10 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package tn.esprit.overpowered.byusforus.entities.candidat;
+package tn.esprit.overpowered.byusforus.entities.users;
 
 import java.io.Serializable;
+import javax.persistence.Column;
 import javax.persistence.DiscriminatorColumn;
 import javax.persistence.DiscriminatorValue;
 import javax.persistence.Entity;
@@ -19,17 +20,15 @@ import javax.validation.constraints.Pattern;
 
 @Entity
 @Inheritance(strategy = InheritanceType.SINGLE_TABLE)
-@DiscriminatorColumn(name = "ObjectType")
+@DiscriminatorColumn(name = "USER_TYPE")
 @DiscriminatorValue(value = "USER")
 public class User implements Serializable{
 
     private static final long serialVersionUID = 1L;
     @Id
-    @GeneratedValue(
-            strategy = GenerationType.IDENTITY
-    )
-    private int id;
-    
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @Column(unique=true)
      @Pattern(regexp="(?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+"
              + "(?:\\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*|\""
              + "(?:[\\x01-\\x08\\x0b\\x0c\\x0e-\\x1f\\x21"
@@ -42,12 +41,47 @@ public class User implements Serializable{
              message="{invalid.email}")
     private String email;
     
-     
+    @Column(unique=true)
     private String username;
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public byte[] getSalt() {
+        return salt;
+    }
+
+    public void setSalt(byte[] salt) {
+        this.salt = salt;
+    }
+
+    public byte[] getPassword() {
+        return password;
+    }
+
+    public void setPassword(byte[] password) {
+        this.password = password;
+    }
     
+    @Column(columnDefinition="BINARY(32) NOT NULL")
+    private byte[] salt;
     
-    private String password;
-    public int getId() {
+    @Column(columnDefinition="BINARY(32) NOT NULL")
+    private byte[] password;
+    public Long getId() {
         return this.id;
     }
