+++ piJEE/piJEE-ejb/src/main/java/tn/esprit/overpowered/byusforus/services/authentication/AuthenticationFacadeRemote.java
@@ -1,0 +1,24 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package tn.esprit.overpowered.byusforus.services.authentication;
+
+import java.security.NoSuchAlgorithmException;
+import javax.ejb.Remote;
+
+/**
+ *
+ * @author EliteBook
+ */
+@Remote
+public interface AuthenticationFacadeRemote {
+
+    String login(String username, String password) throws NoSuchAlgorithmException;
+
+    int logout();
+
+    int isLoggedIn(final String token);
+    
+}
