+++ piJEE/piJEE-ejb/src/main/java/tn/esprit/overpowered/byusforus/services/candidat/CandidateFacade.java
@@ -9,7 +9,7 @@ import tn.esprit.overpowered.byusforus.entities.util.AbstractFacade;
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
-import tn.esprit.overpowered.byusforus.entities.candidat.Candidate;
+import tn.esprit.overpowered.byusforus.entities.users.Candidate;
 
 /**
  *
