+++ piJEE/piJEE-ejb/src/main/java/tn/esprit/overpowered/byusforus/services/entrepriseprofile/CompanyAdminFacadeRemote.java
@@ -7,7 +7,7 @@ package tn.esprit.overpowered.byusforus.services.entrepriseprofile;
 
 import java.util.List;
 import javax.ejb.Remote;
-import tn.esprit.overpowered.byusforus.entities.entrepriseprofile.CompanyAdmin;
+import tn.esprit.overpowered.byusforus.entities.users.CompanyAdmin;
 
 /**
  *
