+++ piJEE/piJEE-ejb/src/main/java/tn/esprit/overpowered/byusforus/services/entrepriseprofile/EmployeeFacadeLocal.java
@@ -7,7 +7,7 @@ package tn.esprit.overpowered.byusforus.services.entrepriseprofile;
 
 import java.util.List;
 import javax.ejb.Local;
-import tn.esprit.overpowered.byusforus.entities.entrepriseprofile.Employee;
+import tn.esprit.overpowered.byusforus.entities.users.Employee;
 
 /**
  *
