+++ piJEE/piJEE-ejb/src/main/java/tn/esprit/overpowered/byusforus/services/users/UserFacade.java
@@ -3,13 +3,13 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package tn.esprit.overpowered.byusforus.services.candidat;
+package tn.esprit.overpowered.byusforus.services.users;
 
 import tn.esprit.overpowered.byusforus.entities.util.AbstractFacade;
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
-import tn.esprit.overpowered.byusforus.entities.candidat.User;
+import tn.esprit.overpowered.byusforus.entities.users.User;
 
 /**
  *
